+++ src/Action/Route/RouteOverview.php
@@ -4,11 +4,15 @@ declare(strict_types=1);
 
 namespace Heptacom\HeptaConnect\Storage\ShopwareDal\Action\Route;
 
+use Doctrine\DBAL\Connection;
+use Heptacom\HeptaConnect\Dataset\Base\ClassStringReferenceCollection;
 use Heptacom\HeptaConnect\Dataset\Base\UnsafeClassString;
+use Heptacom\HeptaConnect\Portal\Base\StorageKey\PortalNodeKeyCollection;
 use Heptacom\HeptaConnect\Storage\Base\Action\Route\Overview\RouteOverviewCriteria;
 use Heptacom\HeptaConnect\Storage\Base\Action\Route\Overview\RouteOverviewResult;
 use Heptacom\HeptaConnect\Storage\Base\Contract\Action\Route\RouteOverviewActionInterface;
 use Heptacom\HeptaConnect\Storage\Base\Exception\InvalidOverviewCriteriaException;
+use Heptacom\HeptaConnect\Storage\Base\Exception\UnsupportedStorageKeyException;
 use Heptacom\HeptaConnect\Storage\ShopwareDal\StorageKey\PortalNodeStorageKey;
 use Heptacom\HeptaConnect\Storage\ShopwareDal\StorageKey\RouteStorageKey;
 use Heptacom\HeptaConnect\Storage\ShopwareDal\Support\DateTime;
@@ -32,6 +36,59 @@ final class RouteOverview implements RouteOverviewActionInterface
     public function overview(RouteOverviewCriteria $criteria): iterable
     {
         $builder = $this->getBuilderCached();
+        $capabilityFilter = $criteria->getCapabilityFilter();
+
+        if ($capabilityFilter !== null) {
+            $builder->andWhere($builder->expr()->in('capability.name', ':caps'));
+            $builder->setParameter('caps', $capabilityFilter, Connection::PARAM_STR_ARRAY);
+        }
+
+        $portalNodeKeys = $criteria->getSourcePortalNodeKeyFilter();
+
+        if ($portalNodeKeys instanceof PortalNodeKeyCollection) {
+            $portalNodeIds = [];
+
+            foreach ($portalNodeKeys as $portalNodeKey) {
+                if (!$portalNodeKey instanceof PortalNodeStorageKey) {
+                    throw new UnsupportedStorageKeyException(\get_class($portalNodeKey));
+                }
+
+                $portalNodeIds[] = $portalNodeKey->getUuid();
+            }
+
+            $builder->andWhere($builder->expr()->in('source_portal_node.id', ':sourcePortals'));
+            $builder->setParameter('sourcePortals', Id::toBinaryList($portalNodeIds), Connection::PARAM_STR_ARRAY);
+        }
+
+        $portalNodeKeys = $criteria->getTargetPortalNodeKeyFilter();
+
+        if ($portalNodeKeys instanceof PortalNodeKeyCollection) {
+            $portalNodeIds = [];
+
+            foreach ($portalNodeKeys as $portalNodeKey) {
+                if (!$portalNodeKey instanceof PortalNodeStorageKey) {
+                    throw new UnsupportedStorageKeyException(\get_class($portalNodeKey));
+                }
+
+                $portalNodeIds[] = $portalNodeKey->getUuid();
+            }
+
+            $builder->andWhere($builder->expr()->in('target_portal_node.id', ':targetPortals'));
+            $builder->setParameter('targetPortals', Id::toBinaryList($portalNodeIds), Connection::PARAM_STR_ARRAY);
+        }
+
+        $entityTypes = $criteria->getEntityTypeFilter();
+
+        if ($entityTypes instanceof ClassStringReferenceCollection) {
+            $entities = [];
+
+            foreach ($entityTypes as $entityType) {
+                $entities[] = (string) $entityType;
+            }
+
+            $builder->andWhere($builder->expr()->in('entity_type.type', ':entities'));
+            $builder->setParameter('entities', $entities, Connection::PARAM_STR_ARRAY);
+        }
 
         foreach ($criteria->getSort() as $field => $direction) {
             $dbalDirection = $direction === RouteOverviewCriteria::SORT_ASC ? 'ASC' : 'DESC';